[Version 0.1.0] Initial version of RetVec ported over for external, open-source release. (#2)
Initial version of RetVec's open-source release for TensorFlow. This CL contains the RetVec layers (including the Binarizer, Integerizer, and Embedding layers) as well as two REW* word models (REWFormer and REWCNN), all fully unit-tested. It includes the pretraining script for REW* models and the default base configs to train REW* models as well.